Move controller select menu stuff into its own file
