+++ src/game/area.h
@@ -135,30 +135,6 @@ struct WarpTransition {
     /*0x04*/ struct WarpTransitionData data;
 };
 
-struct ButtonName {
-    /*0x00*/ u8 pad[2];
-    /*0x02*/ u16 mask;          // Button mask.
-    /*0x04*/ const char* name;  // Button name string.
-}; /*0x08*/
-
-typedef union {
-    struct PACKED {
-        /*0x00*/ u16 mask;          // Button mask.
-        /*0x02*/ u8 x : 5, w : 3;   // [0..31], [0..7]
-        /*0x03*/ u8 y : 5, h : 3;   // [0..31], [0..7]
-    }; /*0x04*/
-    u32 raw;
-} ButtonHighlight; /*0x04*/
-
-struct ControllerIcon {
-    /*0x00*/ u8 pad[2];
-    /*0x02*/ u16 type;              // Controller type.
-    /*0x04*/ Texture* texture;      // Pointer to the controller texture for the above type.
-#ifdef CONTROLLERS_INPUT_DISPLAY
-    /*0x08*/ const ButtonHighlight (*buttonHighlightList)[];  // Pointer to the list of button highlights for the above texture.
-#endif // CONTROLLERS_INPUT_DISPLAY
-}; /*0x0C*/
-
 enum CurrSaveFileNum {
     SAVE_FILE_NUM_A = 0x1,
     SAVE_FILE_NUM_B,
